Enabled removal of category
